jwt는 옳은데 API키가 잘못된 경우 테스팅 추가 및, int 형변환을 더 안전한 long 형변환으로 변경

+++ src/main/java/com/rest1/domain/member/member/service/MemberService.java
@@ -61,7 +61,7 @@ public class MemberService {
         return  authTokenService.payloadOrNull(accessToken);
     }
 
-    public Optional<Member> findById(int id) {
+    public Optional<Member> findById(long id) {
         return memberRepository.findById((long)id);
     }
